init_db/requirements.txt
# Stock Data Loader Dependencies
yfinance>=0.2.64
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas>=2.0.0
requests>=2.31.0 